Add files via upload
In the ever-evolving landscape of social media, understanding and predicting the reach of content is crucial for effective digital marketing strategies. This project focuses on the harnessing the powering of machine learning to analyze and predict the Instagram reach, providing marketers and content creators with valuable insights. Key components of this project include data preprocessing to handle various types of Instagram metrices, feature engineering models for prediction. The three algorithms are used to predict the accuracy of the prediction and to find the algorithm which gives the best accuracy. The algorithms are Passive Aggressive Regressor, Random Forest Regressor and Gradient Boosting Regressor.
Data Collection - The dataset of 120 records used for the analysis. It includes follows, saves, comments, shares, likes and so on.
Data Preprocessing - Ensure the data's are in the suitable format for the analysis. It includes the steps of Data cleaning, Data Transformation, Feature scaling and Feature selection.
Data Visualization - By using the python library matplotlib, seaborn and plotly to convert the data into visual for the better understanding of the data.
Model Training &Evaluation - Split the data into train and test, like 70% for training and 30% for testing and fit that into the model for finding its accuracy.
Prediction - Once the model was trained and evaluated, compare the three models and find the one which gives the best accuracy.
Result - The model Passive Aggressive Regressor gives the best accuracy.